% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_M.R
\name{comp_M}
\alias{comp_M}
\title{Compute the Adjusted Racial Margin of Victory}
\usage{
comp_M(Vm, Vw)
}
\arguments{
\item{Vm}{Numeric vector containing the top minority candidate's vote share,
expressed from 0 to 100.}

\item{Vw}{Numeric vector containing the top White candidate's vote share,
expressed from 0 to 100.}
}
\value{
A numeric vector of adjusted racial margins of victory. Inputs of
  length 1 are expanded to the length of the other input.
}
\description{
Computes the adjusted racial margin of victory from the vote shares of the
top minority and White candidates.
}
\examples{
top_minority <- c(18, 40, 85, 20)
top_white <- c(60, 40, 10, 34)

M_vec_obs <- comp_M(
  Vm = top_minority,
  Vw = top_white
)
minorep(M = M_vec_obs, C = c(50, 45, 65, 35))
}
\references{
Atsusaka, Y. (2021). A Logical Model for Predicting Minority
  Representation: Application to Redistricting and Voting Rights Cases.
  \emph{American Political Science Review}, 115(4), 1210-1225.
  \doi{10.1017/S000305542100054X}
}
