% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emerge.R
\name{emerge}
\alias{emerge}
\title{emerge}
\usage{
emerge(M, C, sd = 1)
}
\arguments{
\item{M}{a vector of raw margin of victories}

\item{C}{a vector of the percentages of minority voters in districts}

\item{sd}{the pre-specified value of the standard deviation in the standard normal CDF}
}
\value{
A list containing main results ($Results) and related statistics ($Stats).
}
\description{
\code{emerge} computes a vector of probabilities that minority candidate emerges given M and C
}
\examples{
rmargin = c(20, 50, 30)
percent = c(40, 70, 85)
emerge_prob(M=rmargin, C=percent)
emerge_prob(M=rmargin, C=percent, sd=5)
}
