% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logical-package.R
\docType{package}
\name{logical-package}
\alias{logical}
\alias{logical-package}
\title{logical: Predictions from a Logical Model of Minority Representation}
\description{
The logical package computes and visualizes quantitative predictions from a
logical model of minority candidate emergence and electoral success. Its
functions support district-level predictions, jurisdiction-level simulation,
and redistricting scenario analysis.
}
\references{
Atsusaka, Y. (2021). A Logical Model for Predicting Minority
  Representation: Application to Redistricting and Voting Rights Cases.
  \emph{American Political Science Review}, 115(4), 1210-1225.
  \doi{10.1017/S000305542100054X}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/YukiAtsusaka/logical}
  \item Report bugs at \url{https://github.com/YukiAtsusaka/logical/issues}
}

}
\author{
\strong{Maintainer}: Yuki Atsusaka \email{yuki.atsusaka@gmail.com}

}
\keyword{internal}
