% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/racialMV.R
\name{margin}
\alias{margin}
\title{margin}
\usage{
margin(Vm, Vw)
}
\arguments{
\item{Vm}{a vector of the vote share (in %) for the top minority candidate}

\item{Vw}{a vector of the vote share (in %) for the top White candidate}
}
\value{
A list containing the racial margin of victory
}
\description{
\code{margin} computes the racial margin of victory given the vote shares for the top minority and White candidates
}
\examples{
topBlack = c(18, 40, 85, 20)
topWhite = c(60, 40, 10, 34)
m <- margin(Vm=topBlack, Vw=topWhite)
hist(m)
}
