% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.minorep.R
\name{plot.minorep}
\alias{plot.minorep}
\title{plot.minorep}
\usage{
\method{plot}{minorep}(M, C, pch = NULL)
}
\arguments{
\item{M}{a vector of raw margin of victories}

\item{C}{a vector of the percentages of minority voters in districts}
}
\value{
A list containing main results ($Results) and related statistics ($Stats).
}
\description{
\code{plot.minorep} visualizes the probability of minority candidate emergence and electoral success with given M and C
}
\examples{
rmargin = c(20, 50, 30)
VAP = c(40, 70, 85)
plot.minorep(M=rmargin, C=VAP)
}
