% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.redistrict.R
\name{plot.redistrict}
\alias{plot.redistrict}
\title{plot.redistrict}
\usage{
\method{plot}{redistrict}(plans, range)
}
\arguments{
\item{M}{a vector of raw margin of victories}

\item{C}{a vector of the percentages of minority voters in districts}
}
\value{
A list containing main results ($Results) and related statistics ($Stats).
}
\description{
\code{plot.redistrict} visualizes the impact of redistricting on minority descriptive representation
}
\examples{
plan1 <- redistrict(coethnic=1, crossover=0)
plan2 <- redistrict(coethnic=1. crossover=0.3)
myplans = cbind(plan1, plan2)
mtrange = c(44,55) # From 44\% to 55\%
plot.redistrict(plans=myplans, range=myrange)
 text(x=start, y=1.1, labels="Moderate white crossover",
     cex=1, col="maroon", font=2)
text(x=start+10, y=-0.09, labels="No white crossover",
     cex=1, col="seagreen", font=2)
title("With Strong Minority Bloc Voting")
}
