% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_minorep.R
\name{plot_minorep}
\alias{plot_minorep}
\title{plot_minorep}
\usage{
plot_minorep(M, C, pch = NULL)
}
\arguments{
\item{M}{a vector of raw margin of victories}

\item{C}{a vector of the percentages of minority voters in districts}
}
\value{
A list containing main results ($Results) and related statistics ($Stats).
}
\description{
\code{plot_minorep} visualizes the probability of minority candidate emergence and electoral success with given M and C
}
\examples{
M_vec = c(20, 50, 30) # Racial margin of victories from three districts
C_vec = c(40, 70, 85) # Percentage of minority voters from the same districts
plot_minorep(M=rmargin, C=VAP) # Assuming no turnout gap
}
