% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redistrict.R
\name{redistrict}
\alias{redistrict}
\title{redistrict}
\usage{
redistrict(coethnic, crossover, sd)
}
\arguments{
\item{coethnic}{the level of minority co-ethnic voting}

\item{crossover}{the level of White crossover voting}
}
\value{
A list containing main results ($Results) and related statistics ($Stats).
}
\description{
\code{redistrict} simulates the probability of minority candidate emergence when changing the district racial composition
}
\examples{
out <- redistrict(coethnic=0.9, crossover=0.3, sd=5)
out2 <- redistrict(coethnic=0.9, crossover=0.6, sd=5)
plot(0, type="n", ylim=c(0,1.1),xlim=c(45,65),
    ylab="Pr(Minority Candidate Emergence)",xlab="C (\% of Minority Voters)",
    mgp=c(2,0.7,0))
lines(P ~ C.tilde, col="seagreen",lwd=2)
lines(P ~ C.tilde, col="maroon",lwd=2)
points(x=50, y=out[50], pch=16, cex=2, col=alpha("seagreen",0.9))
points(x=60, y=out[60], pch=16, cex=2, col=alpha("seagreen",0.9))
text(x=50, y=out[50]-0.09, labels=round(out[50],d=3), col="seagreen")
text(x=60, y=out[60]-0.09, labels=round(out[60],d=3), col="seagreen")
points(x=50, y=out2[50], pch=16, cex=2, col=alpha("maroon", 0.9))
points(x=60, y=out2[60], pch=16, cex=2, col=alpha("maroon", 0.9))
text(x=50, y=out2[50]+0.09, labels=round(out2[50],d=3), col="maroon")
text(x=60, y=out2[60]+0.09, labels=round(out2[60],d=3), col="maroon")
}
