% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_redistrict.R
\name{sim_redistrict}
\alias{sim_redistrict}
\title{Simulate Redistricting Predictions}
\usage{
sim_redistrict(coethnic, crossover, gap = NULL)
}
\arguments{
\item{coethnic}{Numeric value from 0 to 1 giving the proportion of minority
voters expected to support the minority candidate.}

\item{crossover}{Numeric value from 0 to 1 giving the proportion of White
voters expected to support the minority candidate.}

\item{gap}{Optional length-two numeric vector containing minority and White
turnout rates, each expressed from 0 to 1.}
}
\value{
A numeric vector of 991 predicted probabilities corresponding to
  minority electorate percentages from 1 to 100 in increments of 0.1.
}
\description{
Simulates minority candidate emergence and electoral success as district
racial composition varies from 1 to 100 percent.
}
\examples{
sim1 <- sim_redistrict(coethnic = 0.9, crossover = 0)
sim2 <- sim_redistrict(coethnic = 0.9, crossover = 0.3)

start <- 45
end <- 55
plot_redistrict(cbind(sim1, sim2), range = c(start, end))
}
\references{
Atsusaka, Y. (2021). A Logical Model for Predicting Minority
  Representation: Application to Redistricting and Voting Rights Cases.
  \emph{American Political Science Review}, 115(4), 1210-1225.
  \doi{10.1017/S000305542100054X}
}
