% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logic.R
\name{truth_table}
\alias{truth_table}
\title{Create a truth table from logical expressions}
\usage{
truth_table(inputs, ...)
}
\arguments{
\item{inputs}{A character vector of basic inputs}

\item{...}{A set of logical conditions (see \code{\link[stats:Logic]{stats::Logic()}})}
}
\description{
Create a truth table from logical expressions
}
\examples{
truth_table(c("p", "q"), p|q, p&q)

}
