% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_M.R
\name{comp_M}
\alias{comp_M}
\title{comp_M}
\usage{
comp_M(Vm, Vw)
}
\arguments{
\item{Vm}{a vector of the vote share in percent for the top minority candidate}

\item{Vw}{a vector of the vote share in percent for the top White candidate}
}
\value{
A vector of the racial margin of victories based on observed vote shares
}
\description{
\code{comp_M} computes the racial margin of victory given the vote shares for the top minority and White candidates
}
\examples{
top_minority <- c(18, 40, 85, 20) # Top minority candidate's vote share in four districts
top_white <- c(60, 40, 10, 34)    # Top white candidate's vote share in four districts

M_vec_obs <- comp_M(Vm = top_minority, Vw = top_white) # Compute the (adjusted) racial margin of victor
minorep(M = M_vec_obs, C = c(50, 45, 65, 35))
}
