% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_M.R
\name{comp_M}
\alias{comp_M}
\title{comp_M}
\usage{
comp_M(Vm, Vw)
}
\arguments{
\item{Vm}{a vector of the vote share (in %) for the top minority candidate}

\item{Vw}{a vector of the vote share (in %) for the top White candidate}
}
\value{
A vector of the (adjusted) racial margin of victories
}
\description{
\code{comp_M} computes the racial margin of victory given the vote shares for the top minority and White candidates
}
\examples{
top_minority = c(18, 40, 85, 20) # Vote share (in \%)of the top minority candidates in four districts
top_white = c(60, 40, 10, 34)    # Vote share (in \%)of the top White candidates in the same districts
m <- comp_M(Vm=top_minority, Vw=top_white)
hist(m)
}
