% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logical-package.R
\docType{package}
\name{logical-package}
\alias{logical}
\alias{logical-package}
\title{logical: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
