% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/n_minorep.R
\name{n_minorep}
\alias{n_minorep}
\title{n_minorep}
\usage{
n_minorep(model_pred)
}
\arguments{
\item{model_pred}{A vector of model predictions obtained from \code{minorep}}
}
\value{
A vector of the predicted number of minority candidates who run for office and win in given districts
}
\description{
\code{n_minorep} computes a vector of the number of minority candidates who run for office and win in given districts
}
\examples{
M_vec = c(20, 50, 30)
C_vec = c(40, 70, 85)
p_vec <- minorep(M=M_vec, C=C_vec, gap=c(0.5, 0.6)) # Assuming that minority turnout is 0.5 and White turnout is 0.6 
n_minorep(model_pred=p_vec)
}
