% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_M.R
\name{sim_M}
\alias{sim_M}
\title{Simulate the Adjusted Racial Margin of Victory}
\usage{
sim_M(C, coethnic, crossover)
}
\arguments{
\item{C}{Numeric vector of minority electorate percentages, expressed from 0
to 100.}

\item{coethnic}{Numeric vector containing the proportion of minority voters
expected to support the minority candidate, expressed from 0 to 1.}

\item{crossover}{Numeric vector containing the proportion of White voters
expected to support the minority candidate, expressed from 0 to 1.}
}
\value{
A numeric vector of simulated adjusted racial margins of victory.
  Inputs of length 1 are expanded to the common input length.
}
\description{
Computes adjusted racial margins of victory from minority electorate shares
and expected coethnic and crossover voting.
}
\examples{
# Simulate M from substantive assumptions
C_hypothetical <- c(40, 50, 60)
bloc <- 1        
cross <- 0.3     

M_vec_sim <- sim_M(C = C_hypothetical, coethnic = bloc, crossover = cross)
M_vec_sim

minorep(M = M_vec_sim, C = C_hypothetical)
}
\references{
Atsusaka, Y. (2021). A Logical Model for Predicting Minority
  Representation: Application to Redistricting and Voting Rights Cases.
  \emph{American Political Science Review}, 115(4), 1210-1225.
  \doi{10.1017/S000305542100054X}
}
