% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_M.R
\name{sim_M}
\alias{sim_M}
\title{sim_M}
\usage{
sim_M(C, coethnic, crossover)
}
\arguments{
\item{C}{a scalar denoting the (turnout adjusted) percentages of minority voters in a district of interest}

\item{coethnic}{a scaler denoting the proportion of minority voters who would vote for the minority candidate in biracial elections with a single minority and White candidates.}

\item{crossover}{a scaler denoting the proportion of White voters who would vote for the minority candidate in biracial elections with a single minority and White candidates.}
}
\value{
A vector of the racial margin of victories based on substantive knowledge on the levels of minority bloc and crossover voting
}
\description{
\code{sim_M} simulates the values of the (adjusted) racial margin of victory based on the expected level of coethnic and crossover voting
}
\examples{
Simulating M from substantive knowledge
C_hypothetical <- c(40,50,60)  # Hypothetical percentages of \% minority voters
bloc <- 1        # Proportion of minority voters who vote for a single (hypothetical) minority candidate
cross <- 0.3     # Proportion of white voters who vote for a single (hypothetical) minority candidate

M_vec_sim <- sim_M(C = C_hypothetical, coethnic = bloc, crossover = cross)
M_vec_sim

minorep(M = M_vec_sim, C = C_hypothetical)
}
