% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_M.R
\name{sim_M}
\alias{sim_M}
\title{sim_M}
\usage{
sim_M(C, coethnic, crossover)
}
\arguments{
\item{C}{a scalar denoting the (turnout adjusted) percentages of minority voters in a district of interest}

\item{coethnic}{a scaler denoting the proportion of minority voters who would vote for the minority candidate in biracial elections with a single minority and White candidates.}

\item{crossover}{a scaler denoting the proportion of White voters who would vote for the minority candidate in biracial elections with a single minority and White candidates.#' @return A vector of predicted probabilities that minority candidates run for office and win in given districts (defined by M and C)}
}
\description{
\code{sim_M} simulates the values of the (adjusted) racial margin of victory based on the expected level of coethnic and crossover voting
}
\examples{
sim_M(C=40, coethnic=1, crossover=0.3) 
}
